Fix "dir" in GJKDistance function (#290)
Replace the overloaded dir variable with a new variable closest_pt.
The variable dir is overloaded in GJKDistance function, that it means both the point within the simplex
 that is closest to the origin, and the directional vector along which to compute the support mapping. 
One issue caused by this overload, is that when calling extractClosestPoint, we should pass in the 
closest point, but instead passed in the directional vector. In this PR I introduce a separate variable 
closest_p to represent the closest point.