Literal defined as ccd_real_t
A literal in a type-deduced function call was declared as a double literal
(e.g., `1.`). When ccd_real_t is a double, this is fine. But when it is
a float, the type cannot be successfully deduced.

We correct this by explicitly calling it out as a ccd_real_t-typed value.