Report configuration state for narrowphase errors (#381)
This allows low-level, narrowphase algorithms to indicate to higher-level
callers that something went so wrong that it would be helpful to capture
the configuration that led to the error.

Add exception
  1. Add exception type.
  2. Instrument some know problem call sites with the new exception.
  3. Expand unit tests to cover the exceptions.
     i. Add FCL_EXPECT_THROWS_MESSAGE* functionality for tests.

Add exception catching
  1. Add exception processing at an appropriate call site.
     1. Requires printing out shapes and solver data; so streaming operator
        support has been added to all corresponding parties.
     2. Added *only* to signed distance functions of both solvers.

Note that `CollisionGeometry` can span mulitple types and this doesn't
necessarily support all types. However, unit tests pass and this can be
extended to encompass more geometry types as necessary.