generateCoordinateSystem() overhaul
Use Eigen operations to calculate generateCoordinateSystem() for an
axis system array instead of the old custom implementation. Make the
implementation correct and complete.

Before, maxCoeff() was being used incorrectly (maxCoeff returns the
highest value in the vector, not the coefficient index).

Provide documentation for getCoordinateSystem and make the argument a
pointer to follow the google style guide.

Also, remove all the unused versions of getCoordinateSystem.