octree: use log odds for free/occupied
Previously, the probability was used as a threshold for when FCL
considered a node free or occupied. However, the default was a
probability of 0.0, which is impossible to represent in an occupancy
octree (since it stores the log-odds, and log(0) is mathematically
undefined). So, instead, have a default free log-odds threshold of
0.0, which indicates 0.5 probability. This matches well the defaults
of octomap, as any non-existing cell is treated as if it has a
probability of 0.5 (log-odds of 0).  Therefore calling
integrateMiss() or updateNode(key, false) on a previously
non-existing node results in a probability less than .5 (so a
log-odds less than 0).

For more info on log-odds, read the paper "OctoMap: An Efficient
Probabilistic 3D Mapping Framework Based on Octrees"

Included in this commit is a new test which fails without the change
to protect from free-space from ever becoming broken in the future.