narrowphase distance: fix signed distance when touching
In the case of mesh or octree distance checks, we do a collision
after distance checking when doing signed distance to find
penetration depth. However, the maximum penetration depth was
initialized to numeric limits min, which is slightly greater than
zero. If the only contact is exactly 0.0 depth, the code fails.
Fix this to the negative max.