Removing unused/unusable code
In a related issue (421) some code in the hierarchy tree was found to be
impossible. Specifically, it could not reasonably be expected to compile.
In fact, it couldn't. But it didn't cause problems because the functions
weren't actually exercised. Because the functions were templated, the
compiler just ignored them.

Given that no one is calling them, they clearly serve no purpose. So, they
are being cut out rather than fixed.