Fixing capsule-capsule distance (#436)
* Fixing capsule-capsule distance

1. Correct definition of capsule definition - i.e. *centered* on the
   capsule origin.
2. Handle the case where the capsule center lines intersect.
3. Make more robust tests which exercise all (hopefully) the corners of
   the code.

* All numerical literals explicitly typed as S

Most of the literals in the test a numbers that are perfectly represented
by both 32- and 64-bit floats.  There are two that cannot be (0.6 and 0.3).
The VisualStudio compiler complained that those literals required a
narrowing conversion when S was a float.

Instead of simply explicitly declaring those two values to be of type S,
all literal numerical values have been cast as such. This is already a
requirement of GTEST (e.g., EXEPCT_EQ(value_of_type_S, 0) doesn't generally
work and needed to be EXPECT_EQ(value_of_type_S, S())). So, it seemed
saner to just be consistent everywhere rather than expect the reader to
infer the patterns.

* Eigen alignment crash on windows