fix doxygen to examine conditional config includes
Some includes only contain code when prerequisite libraries exist.
For example, if FCL_HAVE_OCTOMAP is not nonzero, the OcTree class is
not declared. These preprocessor macros are defined in the generated
fcl/config.h. Doxygen was not configured to search for inclues in the
generated include directory. Therefore Doxygen would never add
documentation for OcTree.

Fix this by adding the generated include path to
DOXYGEN_INCLUDE_PATH, and actually have the Doxyfile.in make use of
the DOXYGEN_INCLUDE_PATH by providing it as an environment variable
to the doxygen tool.