Make constants::eps and variants constexpr
In the constants, pi() and phi() were declared constexpr but eps and its
variants were not. This was an oversight; there is no reason they couldn't
be declared constexpr.

Furthermore, call sites have been updated to clearly state that the values
are constexpr so future developers can easily recognize that no runtime
cost is paid to acquire those values of epsilon.