Make constants::eps constexpr
In the constants, pi() and phi() were declared constexpr but eps and its
variants were not. There is no reason that eps() cannot be constexpr.
However, the epsNN() methods cannot be; they make use of `pow`. And pow
is not declared constexpr across all compilers used in CI. So, for now,
only eps() is constexpr.

Furthermore, call sites have been updated to clearly state that the values
are constexpr so future developers can easily recognize that no runtime
cost is paid to acquire those values of epsilon.