Update EPA-GJK failure messages
The EPA-GJK failure message attempts to dump sufficient state to reproduce
the configuration that produced the error. However, when one of the
participating shapes is Convex, the information was insufficient; we need
to be able to reconstruct the Convex.

Defining operator<< to dump an entire constructor invocation for Convex
seems unreasonable. We ape python's repr() API and add a Representation()
method to all of the participating shapes.

This also changes the formatting of the poses to smooth the process of
creating workable code from the configuration data (comma-separated
matrix).