Limit the circumstances in which CI is kicked off
With the previous spelling, if a forked version of FCL had actions
enabled, then every push to every branch would kick off a mac/ubuntu
CI event. This is ridiculous and telling forkers to not allow github
actions is not the solution.

So, we take two strategies to spare the FCL contributors:

 1. We limit the CI to operate only on the master branch.
 2. The CI tasks become no-ops unless it is run on the FCL repo. This will
    still cause actions to be triggered, but they will silently do
    nothing in zero time.